Complete Observable after finding the first episode matching the clicked ID
This is an optimisation over the existing implementation (which takes O(n) time) - after the improvement it is still O(n), but should now be twice as fast on average. However I would still opt for going with a HashMap which gives us O(1) access times and no overhead building an Observable chain. Surely constructing the HashMap is a bit more expensive than an ArrayList, but that is done just once after loading.